renamed 'minimal' to baseline
